+++ tweetshake/ViewController.swift
@@ -11,7 +11,7 @@ import TwitterKit
 
 class ViewController: UIViewController {
 
-    var lastComposeResult = "not started"
+    var isShowingComposer = false
 
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -24,18 +24,14 @@ class ViewController: UIViewController {
     }
 
     override func motionEnded(motion: UIEventSubtype, withEvent event: UIEvent) {
-        if motion == UIEventSubtype.MotionShake {
+        if motion == UIEventSubtype.MotionShake && !isShowingComposer {
+            isShowingComposer = true
             let composer = TWTRComposer()
 
             composer.setText("I just shook my phone")
 
             composer.showWithCompletion { [weak self] (result) -> Void in
-                if (result == TWTRComposerResult.Cancelled) {
-                    self?.lastComposeResult = "cancelled"
-                }
-                else {
-                    self?.lastComposeResult = "sent"
-                }
+                self?.isShowingComposer = false
             }
         }
     }